Move Status#project to its own controller
The action `Status#project` was moved to the controller
`StatusProject#show`
