+++ src/api/spec/cassettes/StatusProjectController/GET_show/with_groups/1_1_6_1.yml
@@ -1,0 +1,143 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: http://backend:5352/source/Apache/apache2?view=info
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+      User-Agent:
+      - Ruby
+  response:
+    status:
+      code: 404
+      message: project 'Apache' does not exist
+    headers:
+      Content-Type:
+      - text/xml
+      Cache-Control:
+      - no-cache
+      Connection:
+      - close
+      Content-Length:
+      - '140'
+    body:
+      encoding: UTF-8
+      string: |
+        <status code="404">
+          <summary>project 'Apache' does not exist</summary>
+          <details>404 project 'Apache' does not exist</details>
+        </status>
+    http_version: 
+  recorded_at: Thu, 06 Sep 2018 09:16:22 GMT
+- request:
+    method: get
+    uri: http://backend:5352/source/Apache/apache2
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+      User-Agent:
+      - Ruby
+  response:
+    status:
+      code: 404
+      message: project 'Apache' does not exist
+    headers:
+      Content-Type:
+      - text/xml
+      Cache-Control:
+      - no-cache
+      Connection:
+      - close
+      Content-Length:
+      - '140'
+    body:
+      encoding: UTF-8
+      string: |
+        <status code="404">
+          <summary>project 'Apache' does not exist</summary>
+          <details>404 project 'Apache' does not exist</details>
+        </status>
+    http_version: 
+  recorded_at: Thu, 06 Sep 2018 09:16:22 GMT
+- request:
+    method: post
+    uri: http://backend:5352/source/Apache/apache2?cmd=diff&linkrev=base&onlyissues=1&orev=0&view=xml
+    body:
+      encoding: UTF-8
+      string: ''
+    headers:
+      Content-Type:
+      - application/octet-stream
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+      User-Agent:
+      - Ruby
+  response:
+    status:
+      code: 404
+      message: project 'Apache' does not exist
+    headers:
+      Content-Type:
+      - text/xml
+      Cache-Control:
+      - no-cache
+      Connection:
+      - close
+      Content-Length:
+      - '140'
+    body:
+      encoding: UTF-8
+      string: |
+        <status code="404">
+          <summary>project 'Apache' does not exist</summary>
+          <details>404 project 'Apache' does not exist</details>
+        </status>
+    http_version: 
+  recorded_at: Thu, 06 Sep 2018 09:16:22 GMT
+- request:
+    method: get
+    uri: http://backend:5352/getprojpack?expandedrepos=1&nopackages=1&project=Apache&withrepos=1
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+      User-Agent:
+      - Ruby
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Content-Type:
+      - text/xml
+      Cache-Control:
+      - no-cache
+      Connection:
+      - close
+      Content-Length:
+      - '42'
+    body:
+      encoding: UTF-8
+      string: |
+        <projpack repoid="857298135">
+        </projpack>
+    http_version: 
+  recorded_at: Thu, 06 Sep 2018 09:16:22 GMT
+recorded_with: VCR 4.0.0
